pyproject.toml
[tool.poetry]
name = "deepl-tr-pp"
version = "0.1.3"
description = "deepl translate via pyppeteer"
authors = ["freemt"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/ffreemt/deepl-tr-pyppeteer"
keywords = ["mt", "language"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.7"
pyppeteer2 = "^0.2.2"
logzero = "^1.6.3"
environs = "^9.2.0"
python-dotenv = "^0.15.0"
absl-py = "^0.11.0"
docx2txt = "^0.8"
pydantic = "^1.7.3"
pyquery = "^1.4.3"
pyperclip = "^1.8.1"
cchardet = "^2.1.7"
python-docx = "^0.8.10"
tqdm = "^4.56.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.19.0"
flake8 = "^3.8.4"
pyinstaller = "^4.1"

[tool.poetry.scripts]
deepl-tr-pp = "deepl_tr_pp.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
